Remove separate definitions section + remove unit test for heading autopopulation
