+++ cms/core/blocks/section_blocks.py
@@ -1,10 +1,8 @@
 from typing import TYPE_CHECKING, ClassVar
 
 from django.utils.text import slugify
-from django.utils.translation import gettext_lazy as _
-from wagtail.blocks import ListBlock, RichTextBlock, StreamBlock, StructBlock
+from wagtail.blocks import RichTextBlock, StreamBlock, StructBlock
 from wagtail.images.blocks import ImageChooserBlock
-from wagtail.snippets.blocks import SnippetChooserBlock
 from wagtailmath.blocks import MathBlock
 
 from cms.core.blocks import (
@@ -18,10 +16,10 @@ from cms.core.blocks import (
     VideoEmbedBlock,
     WarningPanelBlock,
 )
-from cms.core.models import GlossaryTerm
+from cms.core.blocks.glossary_terms import GlossaryTermsBlock
 
 if TYPE_CHECKING:
-    from wagtail.blocks import StreamValue, StructValue
+    from wagtail.blocks import StructValue
 
 
 class SectionContentBlock(StreamBlock):
@@ -38,34 +36,11 @@ class SectionContentBlock(StreamBlock):
     equation = MathBlock(group="DataVis", icon="decimal")
     ons_embed = ONSEmbedBlock(group="DataVis", label="ONS General Embed")
     related_links = RelatedLinksBlock(icon="link")
+    glossary_terms = GlossaryTermsBlock()
 
     class Meta:
         template = "templates/components/streamfield/stream_block.html"
         block_counts: ClassVar[dict[str, dict]] = {"related_links": {"max_num": 1}}
-
-
-class GlossarySectionBlock(StructBlock):
-    title = HeadingBlock(default=_("Definitions"))
-    content = ListBlock(SnippetChooserBlock(GlossaryTerm))
-
-    class Meta:
-        template = "templates/components/streamfield/glossary_section_block.html"
-
-    def to_table_of_contents_items(self, value: "StructValue") -> list[dict[str, str]]:
-        """Convert the value to the table of contents component macro format."""
-        return [{"url": "#" + slugify(value["title"]), "text": value["title"]}]
-
-    def get_context(self, value: "StreamValue", parent_context: dict | None = None) -> dict:
-        """Inject formatted glossary terms to be used with ONS Accordion component."""
-        context: dict = super().get_context(value, parent_context)
-        context["formatted_glossary_terms"] = [
-            {
-                "title": glossary_term.name,
-                "content": glossary_term.definition,
-            }
-            for glossary_term in value["content"]
-        ]
-        return context
 
 
 class SectionBlock(StructBlock):
