+++ cms/core/blocks/stream_blocks.py
@@ -15,7 +15,7 @@ from cms.core.blocks import (
     VideoEmbedBlock,
     WarningPanelBlock,
 )
-from cms.core.blocks.section_blocks import GlossarySectionBlock, SectionBlock
+from cms.core.blocks.section_blocks import SectionBlock
 
 if TYPE_CHECKING:
     from wagtail.blocks import StreamValue
@@ -25,7 +25,6 @@ class SectionStoryBlock(StreamBlock):
     """The core section StreamField block definition."""
 
     section = SectionBlock()
-    glossary_section = GlossarySectionBlock(label="Definitions")
 
     class Meta:
         template = "templates/components/streamfield/stream_block.html"
