+++ cms/methodology/tests/test_models.py
@@ -6,7 +6,7 @@ from django.test import TestCase, override_settings
 from django.utils.formats import date_format
 from wagtail.test.utils import WagtailTestUtils
 
-from cms.core.tests.factories import ContactDetailsFactory, GlossaryTermFactory
+from cms.core.tests.factories import ContactDetailsFactory
 from cms.methodology.tests.factories import MethodologyPageFactory, MethodologyRelatedPageFactory
 
 
@@ -40,16 +40,6 @@ class MethodologyPageTestCase(WagtailTestUtils, TestCase):
     def test_table_of_contents_without_related_publications(self):
         MethodologyRelatedPageFactory(parent=self.page)
         self.assertIn({"url": "#related-publications", "text": "Related publications"}, self.page.table_of_contents)
-
-    def test_table_of_contents_with_glossary_section(self):
-        glossary_term = GlossaryTermFactory()
-        self.page.content = [
-            {
-                "type": "glossary_section",
-                "value": {"title": "Definitions", "content": [glossary_term.pk]},
-            }
-        ]
-        self.assertIn({"url": "#definitions", "text": "Definitions"}, self.page.table_of_contents)
 
     def test_cite_this_page_is_not_shown_when_unticked(self):
         """Test for the cite this page block not present in the template."""
